api/requirements/development.txt
-r production.txt

black
flake8
httpx
motor-stubs
mypy == 0.961
pytest
pytest-asyncio
